+++ Makefile
@@ -1,4 +1,4 @@
-all: build run
+all: build up
 
 build:
 	docker-compose build
