+++ config.ru
@@ -1,0 +1,3 @@
+require File.dirname(__FILE__) + '/bootstrap.rb'
+
+run App::Base
