+++ controllers/html.rb
@@ -1,0 +1,7 @@
+module App
+  class Base < ::Sinatra::Base
+    get '/' do
+      erb :index
+    end
+  end
+end
