+++ libs/database.rb
