+++ libs/tools.rb
